cleanup visuals, separate overlays
