+++ napari/_vispy/visuals/image.py
@@ -1,5 +1,5 @@
 from vispy.scene.visuals import create_visual_node
 
-from .vendored import ImageVisual
+from ..vendored import ImageVisual
 
 Image = create_visual_node(ImageVisual)
