+++ napari/_vispy/visuals/markers.py
@@ -4,9 +4,6 @@ from vispy.scene.visuals import Markers as BaseMarkers
 # Custom markers class is needed for entering 3D rendering mode when a points
 # layer is invisible and the self._data property is None
 class Markers(BaseMarkers):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
     def _compute_bounds(self, axis, view):
         # This if statement needs to be added to vispy master
         if self._data is None:
