+++ napari/_vispy/visuals/points.py
@@ -1,25 +1,13 @@
 from vispy.scene.visuals import Compound, Line, Text
 
+from .clipping_planes_mixin import ClippingPlanesMixin
 from .markers import Markers
-from .vendored.filters.clipping_planes import PlanesClipper
 
 
-class PointsVisual(Compound):
+class PointsVisual(ClippingPlanesMixin, Compound):
     """
     Compound vispy visual for point visualization
     """
 
     def __init__(self):
-        self._clip_filter = PlanesClipper()
         super().__init__([Markers(), Markers(), Line(), Text()])
-
-        for subv in self._subvisuals:
-            subv.attach(self._clip_filter)
-
-    @property
-    def clipping_planes(self):
-        return self._clip_filter.clipping_planes
-
-    @clipping_planes.setter
-    def clipping_planes(self, value):
-        self._clip_filter.clipping_planes = value
